refactor(proxies): make GValT more generic
Before: base proxies would read and write GValT and Sequence[GValT].
After: base proxies read and write GScalarValT and GSequenceValT.
This is more flexible and will allow for reading/writing something other
than sequences of the base object.
As a consequence, the default implementation for _postprocess_get has to
be moved to the concrete proxies.