feat(grid): add scroll method
`scroll` is a helper method that takes a direction and a step size.
It shifts the whole grid in the opposite direction (scrolling DOWN is
shifting UP...) and calculates the amount by which to shift taking into
account row_height, col_width and buffers (each step is a cell in the
given direction). To avoid artifacts, the cell sizes must be uniforms in
that direction.